requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.15.0

requests~=2.25.0
setuptools~=50.3.2
pandas~=1.1.4
beautifulsoup4~=4.9.3
lxml~=4.6.1
src~=0.1.0
numpy~=1.19.4
scikit-learn~=0.23.2
matplotlib~=3.3.3
seaborn~=0.11.0
scipy~=1.5.3
yaml~=0.2.5
pyyaml~=5.3.1